#13 Removed SockJs functionality from WebSocketServer
Removed a lot of code that is not in use (yet).

Cleaned up a bit print-debug messages.

Implemented ping and pong messages. However a bug was discovered
wherein a DDPServer doesn’t create DDPConnection objects; instead, it
creates WebSocketConnection objects.

Note: in the future, DDP server will be an instance of a SockJs server,
not a WebSocketServer.